adreno_tz: Fix GPU target frequency calculation for high refresh rates
It seems that the black-box GPU frequency selection algorithm is unaware
of high refresh rates, and assumes a refresh rate of 60 Hz for its
target frequencies. This results in frequencies that are too low being
selected when a higher refresh rate, such as 90 Hz, is used.

To remedy this, multiply the GPU's busy time by a ratio of the current
refresh rate to 60 Hz. This makes msm-adreno-tz select the right
frequencies needed to avoid stuttering.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: Id8b2f025ce9f74faa20ecbee42f27d69fd74f7de
Signed-off-by: mukesh22584 <mks22584@gmail.com>
Signed-off-by: chandu078 <chandudyavanapelli03@gmail.com>